Revert "Changed all vectors to doubles"
This reverts commit 8b4cde8497a9ec0316a54e5a2a36bad3719fb848.

+++ geometry/circle.cpp
@@ -9,28 +9,28 @@
 
 class Circle {
 private:
-    glm::dvec3 center;
-    double radius;
+    glm::vec3 center;
+    float radius;
 public:
-  Circle(glm::dvec3 center, double radius): center(center), radius(radius) {}
+  Circle(glm::vec3 center, float radius): center(center), radius(radius) {}
 
   /**
    * calculate the intersection point of a ray with a circle
    * @param  ray (incoming ray)
    */
   Intersection intersect(Ray ray){
-    glm::dvec3 originToCenter = center - ray.origin;
-    double projectionLength = glm::dot(ray.direction, originToCenter);
-    double distToCenterSquared = glm::dot(originToCenter, originToCenter);
-    double distAlongProjSquared = projectionLength*projectionLength;
-    double discriminant = distAlongProjSquared - distToCenterSquared + radius*radius;
+    glm::vec3 originToCenter = center - ray.origin;
+    float projectionLength = glm::dot(ray.direction, originToCenter);
+    float distToCenterSquared = glm::dot(originToCenter, originToCenter);
+    float distAlongProjSquared = projectionLength*projectionLength;
+    float discriminant = distAlongProjSquared - distToCenterSquared + radius*radius;
     if(discriminant < 0)
       return Intersection();
     else {
       //Find the closest intersection point by ray*d = quadratic formula
-      double distanceAlongRayPlus = sqrt(discriminant) - projectionLength;
-      double distanceAlongRayMinus = -1.0f * sqrt(discriminant) - projectionLength;
-      glm::dvec3 intersectionPoint = ray.direction * fmin(distanceAlongRayPlus, distanceAlongRayMinus);
+      float distanceAlongRayPlus = sqrtf(discriminant) - projectionLength;
+      float distanceAlongRayMinus = -1.0f * sqrtf(discriminant) - projectionLength;
+      glm::vec3 intersectionPoint = ray.direction * fminf(distanceAlongRayPlus, distanceAlongRayMinus);
       return Intersection(intersectionPoint, ray, glm::normalize(intersectionPoint-center));
     }
   }
